add BME680 library to platformio.ini
tune BM280 and BME680 configurations to indoor use cases

get temperature, humidity and pressure as floats, report with 1 decimal place of precision

+++ src/bme280_sensor.h
@@ -12,15 +12,15 @@ class BME280_Sensor : public Sensor {
   void begin();
   void handle();
 
-  uint16_t temperature() { _mark_read(); return _temperature; };
-  uint16_t humidity() { _mark_read(); return _humidity; };
-  uint16_t pressure() { _mark_read(); return _pressure; };
-  uint16_t altitude() { _mark_read(); return _altitude; };
+  float temperature() { _mark_read(); return _temperature; };
+  float humidity() { _mark_read(); return _humidity; };
+  float pressure() { _mark_read(); return _pressure; };
+  float altitude() { _mark_read(); return _altitude; };
 
  private:
   Adafruit_BME280 _bme;
-  uint16_t _temperature;
-  uint16_t _humidity;
-  uint16_t _pressure;
-  uint16_t _altitude;
+  float _temperature;
+  float _humidity;
+  float _pressure;
+  float _altitude;
 };
